+++ fsl/scripts/remove_ext.py
@@ -6,22 +6,22 @@
 #
 
 
-import sys
-import warnings
-
+import                   sys
 import fsl.utils.path as fslpath
-
-# See atlasq.py for explanation
-with warnings.catch_warnings():
-    warnings.filterwarnings("ignore", category=FutureWarning)
-    import fsl.data.image as fslimage
 
 
 usage = """Usage: remove_ext <list of image paths to remove extension from>
 """.strip()
 
 
-ALLOWED_EXTENSIONS = fslimage.ALLOWED_EXTENSIONS + ['.mnc', '.mnc.gz']
+# This list is defined in the
+# fsl.data.image class, but are duplicated
+# here for performance (to avoid import of
+# nibabel/numpy/etc).
+exts = ['.nii.gz', '.nii',
+        '.img',    '.hdr',
+        '.img.gz', '.hdr.gz',
+        '.mnc',    '.mnc.gz']
 """List of file extensions that are removed by ``remove_ext``. """
 
 
@@ -40,7 +40,7 @@ def main(argv=None):
     removed = []
 
     for path in argv:
-        removed.append(fslpath.removeExt(path, ALLOWED_EXTENSIONS))
+        removed.append(fslpath.removeExt(path, exts))
 
     print(' '.join(removed))